Rover Camera UI
This is an updated version of the UI designed to work with a camera attached to the rover. The controls have increased in size making them easier to use. The sensors information has now been placed to the side to make room for the camera feed. The UI now features an acrylic material.